Finished up Piece Moves
Move checks are complete for all pieces.

Renamed files and class names to match the purpose of the functions.

En Passant and Castling will be handled within the piece classes. Along with all possible positions for pieces.

Move functions will need to be updated to not display text for moves.